arm32: fix SIGBUS on unaligned IPv4 address access
On 32-bit ARM, a libpcap buffer is 4-byte aligned and the Ethernet
header is 14 bytes, placing the IPv4 header at a 2-byte boundary.
Direct 32-bit loads of ip_src/ip_dst (e.g. via *(uint32_t*)) fault
with SIGBUS because the ARMv7 architecture requires natural alignment
for multi-byte loads.

Three changes fix the runtime crash:

* src/sfip/sf_ip.cc: replace misaligned cast with memcpy when storing
  an IPv4-mapped address into the internal ip32 array.

* src/protocols/ipv4.h: add ARM32-conditional get_src()/get_dst()
  accessors that use __builtin_memcpy so the compiler emits byte-safe
  loads instead of a faulting LDR instruction; route is_src_broadcast()
  and is_dst_broadcast() through those accessors.

* src/codecs/ip/cd_ipv4.cc: replace five direct ip_src/ip_dst reads
  in IP4AddrTests() with get_src()/get_dst() calls.

Three unit-test mocks are also corrected so they compile cleanly on
arm/v7 where size_t is uint32_t and PegCount is uint64_t (both differ
from unsigned long under the ILP32 ABI):

* detector_plugins_mock.h: fix add_host_pattern signature.
* std_connector_test.cc: fix StdConnectorBuffer::acquire parameter type.
* mp_data_bus_test.cc: fix show_stats overload parameter type.

All 162 unit tests pass on linux/arm/v7, linux/arm64, and linux/amd64.

Fixes #458

+++ src/connectors/std_connector/test/std_connector_test.cc
@@ -34,7 +34,7 @@
 StdConnectorBuffer::StdConnectorBuffer(const char*) {}
 StdConnectorBuffer::~StdConnectorBuffer() {}
 void StdConnectorBuffer::start() {}
-Ring2::Writer StdConnectorBuffer::acquire(unsigned long) { return Ring2(0).writer(); }
+Ring2::Writer StdConnectorBuffer::acquire(size_t) { return Ring2(0).writer(); }
 bool StdConnectorBuffer::release(Ring2::Writer const&) { return false; }
 
 using namespace snort;
